requirements.txt
aniso8601==9.0.1
attrs==19.3.0
Automat==0.8.0
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
click==8.0.3
colorama==0.4.3
configobj==5.0.6
constantly==15.1.0
cryptography>=3.2
dbus-python==1.2.16
distro==1.4.0
entrypoints==0.3
Flask==2.0.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
httplib2>=0.19.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==2.0.1
more-itertools==4.2.0
netifaces==0.10.4
numpy==1.21.4
oauthlib==3.1.0
opencv-python==4.5.5.62
pandas==1.3.4
pexpect==4.6.0
Pillow==9.0.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyrsistent==0.15.5
pyserial==3.4
pytesseract==0.3.8
python-dateutil==2.8.2
pytz==2021.3
PyYAML>=5.4
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
ssh-import-id==5.10
Twisted>=20.3.0
wadllib==1.3.3
Werkzeug==2.0.2
zipp==1.0.0
python-docx==0.8.11